refactor(bedrock): rename reasoning-fix to outbound-messages
Consolidate Bedrock's outbound message rewrites under a name that reflects
their role instead of pathology. Contents unchanged.
