+++ packages/inference-providers/src/providers/bedrock/index.ts
@@ -2,4 +2,4 @@
 export { BedrockLangChainModelProvider } from "./langchain.js";
 export { type BedrockProviderOptions } from "./models.js";
 export { discoverAwsProfiles } from "./aws-profiles.js";
-export { stripReasoningForBedrock } from "./reasoning-fix.js";
+export { stripReasoningForBedrock } from "./outbound-messages.js";
